requirements.txt
langchain_groq
selenium
pandas
openpyxl
google-api-python-client
google_auth_oauthlib

celery[redis]
redis
python-dotenv==0.19.0
# gevent==21.8.0
requests==2.26.0
beautifulsoup4==4.10.0  # If needed for scraping